third_party/WebKit/LayoutTests/fast/canvas/toDataURL-noData-expected.txt

  
mimeType: unspecified
PASS: Canvas of size 0x0 created data: url with no content - 'data:,'.
mimeType: unspecified
PASS: Canvas of size 0x100 created data: url with no content - 'data:,'.
mimeType: unspecified
PASS: Canvas of size 100x0 created data: url with no content - 'data:,'.
mimeType: image/jpeg
PASS: Canvas of size 0x0 created data: url with no content - 'data:,'.
mimeType: image/jpeg
PASS: Canvas of size 0x100 created data: url with no content - 'data:,'.
mimeType: image/jpeg
PASS: Canvas of size 100x0 created data: url with no content - 'data:,'.
mimeType: image/webp
PASS: Canvas of size 0x0 created data: url with no content - 'data:,'.
mimeType: image/webp
PASS: Canvas of size 0x100 created data: url with no content - 'data:,'.
mimeType: image/webp
PASS: Canvas of size 100x0 created data: url with no content - 'data:,'.
